# testing-service/docs/swagger.yaml
definitions:
...
  ErrorMessage:
    properties:
      errorDate:
        description: Time of the error, formatted as yyyy-mm-dd hh:mm:ss.mmm.
        example: "2024-05-17 11:33:23.038"
        type: string
      errorMessage:
        description: The error itself, which may be a custom message or an internal
          failure.
        example: Something went wrong.
        type: string
      serviceName:
        description: Name of the service that produced the message.
        example: testing-service
        type: string
      stacktrace:
        description: The stack trace, when one is available; otherwise "No Stacktrace
          Available".
        example: No Stacktrace Available
        type: string
    type: object
  Filter:
    properties:
      condition:
        type: string
      feature:
        type: string
      values:
        items:
          type: string
        type: array
    type: object
...
  SelectionSpecification:
    properties:
      filters:
        items:
          $ref: '#/definitions/Filter'
        type: array
      ids:
        items:
          type: string
        type: array
      searchText:
        type: string
    type: object
...
  TestsRunView:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      errors:
        type: integer
      finish:
        type: string
      id:
        type: string
      start:
        type: string
      status:
        type: string
      testCases:
        type: integer
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
...
info:
  contact: {}
  description: API of the testing service of the Qubership Integration Platform.
  title: Testing Service API
paths:
...
  /api/v1/tests-runs:
    delete:
      consumes:
      - application/json
      operationId: deleteTestsRunsV1
      parameters:
      - description: Tests runs IDs to delete
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Delete tests runs
      tags:
      - V1
      - Tests Runs
    post:
      consumes:
      - application/json
      operationId: findTestsRunsV1
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Sort field
        enum:
        - id
        - start
        - finish
        - status
        - errors
        - test_cases
        - created_by
        - created_at
        in: query
        name: sort_by
        type: string
      - default: ASC
        description: Sort order
        enum:
        - ASC
        - DESC
        in: query
        name: sort_order
        type: string
      - description: Return IDs list
        in: query
        name: return_ids
        type: boolean
      - description: Tests runs specification
        in: body
        name: specification
        required: true
        schema:
          $ref: '#/definitions/SelectionSpecification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/TestsRunView'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Get tests runs
      tags:
      - V1
      - Tests Runs
  /api/v1/tests-runs/{id}:
    delete:
      operationId: deleteTestsRunV1
      parameters:
      - description: Tests run UUID
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Delete tests run
      tags:
      - V1
      - Tests Runs
    get:
      operationId: findTestsRunByIdV1
      parameters:
      - description: Tests run UUID
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TestsRunView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Get tests run
      tags:
      - V1
      - Tests Runs
  /api/v1/tests-runs/{id}/cancel:
    post:
      operationId: cancelTestsRunByIdV1
      parameters:
      - description: Tests run UUID
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Cancel tests run
      tags:
      - V1
      - Tests Runs
  /api/v1/tests-runs/{id}/export:
    post:
      consumes:
      - application/json
      operationId: exportTestsRunV1
      parameters:
      - description: Tests run UUID
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Export tests run
      tags:
      - V1
      - Tests Runs
  /api/v1/tests-runs/cancel:
    post:
      consumes:
      - application/json
      operationId: cancelTestsRunsV1
      parameters:
      - description: Tests runs IDs to cancel
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Cancel tests runs
      tags:
      - V1
      - Tests Runs
  /api/v1/tests-runs/create:
    post:
      consumes:
      - application/json
      operationId: startNewTestsRunV1
      parameters:
      - description: Entities IDs to run corresponding test cases
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      - default: test_cases
        description: Entity type
        enum:
        - test_cases
        - tests_runs
        - test_case_runs
        in: query
        name: from
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created tests run ID
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Start new tests run
      tags:
      - V1
      - Tests Runs
  /api/v1/tests-runs/export:
    post:
      consumes:
      - application/json
      operationId: exportTestsRunsV1
      parameters:
      - description: Tests runs IDs to export
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - text/csv
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorMessage'
      summary: Export tests runs
      tags:
      - V1
      - Tests Runs
produces:
- application/json
swagger: "2.0"
...